src/requirements.txt
pillow==10.2.0
pandas==2.2.1